update to `componentize-py` 0.13.1 (#89)
* update to `componentize-py` 0.13.1

This includes a breaking change: bindings for variant cases are now named using
a `Variant_Case` format instead of the previous `VariantCase` format.  The goal
is to avoid accidental name conflicts when e.g. `VariantCase` is already being
used.

I've bumped the version to v3.0.0 due to that breaking change and also used this
opportunity to version the docs so we can keep links stable.  Now we have v1,
v2, and v3 directories for each set of docs, and we'll aim to keep those links
working when we add v4, etc.

This also includes a fix for reading from and writing to the filesystem via
preopens.

Fixes #88

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* update `spin-sqlite` example to use new API

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>